Include libfec into CMake.
