+++ CMakeLists.txt
@@ -71,6 +71,7 @@ endif ()
 add_subdirectory(clapack-3.2.1-CMAKE)
 add_subdirectory(CBLAS)
 add_subdirectory(plover)
+add_subdirectory(libfec)
 add_subdirectory(src)
 add_subdirectory(docs)
 add_subdirectory(tests)
