+++ libfec/src/CMakeLists.txt
@@ -1,3 +1,3 @@
-set(viterbi27.c fec.c)
-add_library(fec viterbi27.c fec.c)
+set(viterbi27.c)
+add_library(fec viterbi27.c)
