+++ libfec/src/viterbi27.c
@@ -2,94 +2,49 @@
  * Copyright Feb 2004, Phil Karn, KA9Q
  * May be used under the terms of the GNU Lesser General Public License (LGPL)
  */
-#include <stdio.h>
+
 #include <stdlib.h>
-#include <memory.h>
-#include <limits.h>
 #include "fec.h"
 
+void set_viterbi27_polynomial(int polys[2])
+{
+  int state;
 
-typedef union {
-  unsigned int w[64];
-} metric_t;
+  for(state = 0; state < 32; state++) {
+    Branchtab27[0].c[state] = (polys[0] < 0) ^ parity((2*state) & abs(polys[0])) ? 255 : 0;
+    Branchtab27[1].c[state] = (polys[1] < 0) ^ parity((2*state) & abs(polys[1])) ? 255 : 0;
+  }
+}
 
-typedef union {
-  unsigned long w[2];
-} decision_t;
-
-static union branchtab27 {
-  unsigned char c[32];
-} Branchtab27[2] __attribute__ ((aligned(16)));
-
-static int Init = 0;
-
-/* State info for instance of Viterbi decoder
- */
-struct v27 {
-  metric_t metrics1; /* path metric buffer 1 */
-  metric_t metrics2; /* path metric buffer 2 */
-  decision_t *dp;          /* Pointer to current decision */
-  metric_t *old_metrics,*new_metrics; /* Pointers to path metrics, swapped on every bit */
-  decision_t *decisions;   /* Beginning of decisions for block */
-};
-
-/* Initialize Viterbi decoder for start of new frame */
-int init_viterbi27(void *p,int starting_state){
-  struct v27 *vp = p;
+/* Create a new instance of a Viterbi decoder */
+void init_viterbi27(struct v27 *vp, int starting_state)
+{
   int i;
+  int polys[2] = { V27POLYA, V27POLYB };
 
-  if(p == NULL)
-    return -1;
-  for(i=0;i<64;i++)
+  set_viterbi27_polynomial(polys);
+
+  for(i = 0; i < 64; i++)
     vp->metrics1.w[i] = 63;
 
   vp->old_metrics = &vp->metrics1;
   vp->new_metrics = &vp->metrics2;
   vp->dp = vp->decisions;
   vp->old_metrics->w[starting_state & 63] = 0; /* Bias known start state */
-  return 0;
-}
-
-void set_viterbi27_polynomial(int polys[2]){
-  int state;
-
-  for(state=0;state < 32;state++){
-    Branchtab27[0].c[state] = (polys[0] < 0) ^ parity((2*state) & abs(polys[0])) ? 255 : 0;
-    Branchtab27[1].c[state] = (polys[1] < 0) ^ parity((2*state) & abs(polys[1])) ? 255 : 0;
-  }
-  Init++;
-}
-
-/* Create a new instance of a Viterbi decoder */
-void *create_viterbi27(int len){
-  struct v27 *vp;
-
-  if(!Init){
-    int polys[2] = { V27POLYA, V27POLYB };
-    set_viterbi27_polynomial(polys);
-  }
-  if((vp = malloc(sizeof(struct v27))) == NULL)
-     return NULL;
-  if((vp->decisions = malloc((len+6)*sizeof(decision_t))) == NULL){
-    free(vp);
-    return NULL;
-  }
-  init_viterbi27(vp,0);
-
-  return vp;
 }
 
 /* Viterbi chainback */
 int chainback_viterbi27(
-      void *p,
+      struct v27 *vp,
       unsigned char *data, /* Decoded output data */
       unsigned int nbits, /* Number of data bits */
-      unsigned int endstate){ /* Terminal encoder state */
-  struct v27 *vp = p;
+      unsigned int endstate) /* Terminal encoder state */
+{
   decision_t *d;
 
-  if(p == NULL)
+  if(vp == NULL)
     return -1;
+
   d = vp->decisions;
   /* Make room beyond the end of the encoder register so we can
    * accumulate a full byte of decoded data
@@ -102,23 +57,14 @@ int chainback_viterbi27(
    * combine in the cache anyway
    */
   d += 6; /* Look past tail */
-  while(nbits-- != 0){
+  while(nbits-- != 0) {
     int k;
 
     k = (d[nbits].w[(endstate>>2)/32] >> ((endstate>>2)%32)) & 1;
     data[nbits>>3] = endstate = (endstate >> 1) | (k << 7);
   }
+
   return 0;
-}
-
-/* Delete instance of a Viterbi decoder */
-void delete_viterbi27(void *p){
-  struct v27 *vp = p;
-
-  if(vp != NULL){
-    free(vp->decisions);
-    free(vp);
-  }
 }
 
 /* C-language butterfly */
@@ -141,15 +87,17 @@ unsigned int metric,m0,m1,decision;\
  * Note that nbits is the number of decoded data bits, not the number
  * of symbols!
  */
-int update_viterbi27_blk(void *p,unsigned char *syms,int nbits){
-  struct v27 *vp = p;
+int update_viterbi27_blk(struct v27 *vp, unsigned char *syms, int nbits)
+{
   void *tmp;
   decision_t *d;
 
-  if(p == NULL)
+  if(vp == NULL)
     return -1;
+
   d = (decision_t *)vp->dp;
-  while(nbits--){
+
+  while(nbits--) {
     unsigned char sym0,sym1;
 
     d->w[0] = d->w[1] = 0;
@@ -195,6 +143,7 @@ int update_viterbi27_blk(void *p,unsigned char *syms,int nbits){
     vp->old_metrics = vp->new_metrics;
     vp->new_metrics = tmp;
   }
+
   vp->dp = d;
   return 0;
 }
