+++ src/CMakeLists.txt
@@ -47,6 +47,7 @@ add_library(swiftnav-static STATIC ${libswiftnav_SRCS})
 add_dependencies(swiftnav-static generate)
 target_link_libraries(swiftnav-static cblas)
 target_link_libraries(swiftnav-static lapack)
+target_link_libraries(swiftnav-static fec)
 install(TARGETS swiftnav-static DESTINATION lib${LIB_SUFFIX})
 
 if(BUILD_SHARED_LIBS)
@@ -54,6 +55,7 @@ if(BUILD_SHARED_LIBS)
   add_dependencies(swiftnav generate)
   target_link_libraries(swiftnav cblas)
   target_link_libraries(swiftnav lapack)
+  target_link_libraries(swiftnav fec)
   install(TARGETS swiftnav DESTINATION lib${LIB_SUFFIX})
 else(BUILD_SHARED_LIBS)
   message(STATUS "Not building shared libraries")
